Brake out wiseconnect specific code from the common abstraction
